Issue #429 - Reorganize instances - Move ConsoleEffect instance to effectie.instances * Also effectie.instances.future.fromFuture.FromFutureToIdTimeout => effectie.core.FromFuture.FromFutureToIdTimeout
