+++ modules/effectie-cats-effect2/shared/src/test/scala-2/effectie/ce2/fromFutureSpec.scala
@@ -4,10 +4,10 @@ import cats.Id
 import cats.effect._
 import effectie.ce2.compat.CatsEffectIoCompatForFuture
 import effectie.core.FromFuture
+import effectie.core.FromFuture.FromFutureToIdTimeout
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import fromFuture._
-import effectie.instances.future.fromFuture
 import hedgehog._
 import hedgehog.runner._
 
@@ -54,8 +54,8 @@ object fromFutureSpec extends Properties {
         ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
 
       ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
-        implicit val timeout: fromFuture.FromFutureToIdTimeout =
-          fromFuture.FromFutureToIdTimeout(waitFor300Millis.waitFor)
+        implicit val timeout: FromFutureToIdTimeout =
+          FromFutureToIdTimeout(waitFor300Millis.waitFor)
 
         lazy val fa = Future(a)
         val actual  = FromFuture[Id].toEffect(fa)
