+++ modules/effectie-cats-effect3/shared/src/test/scala-3/effectie/ce3/fromFutureSpec.scala
@@ -6,7 +6,7 @@ import cats.effect.unsafe.IORuntime
 import effectie.ce3.compat.CatsEffectIoCompatForFuture
 import effectie.ce3.fromFuture.given
 import effectie.core.FromFuture
-import effectie.instances.future.fromFuture
+import effectie.core.FromFuture.FromFutureToIdTimeout
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
@@ -57,8 +57,8 @@ object fromFutureSpec extends Properties {
         ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
 
       ConcurrentSupport.runAndShutdown(es, WaitFor(300.milliseconds)) {
-        given timeout: fromFuture.FromFutureToIdTimeout =
-          fromFuture.FromFutureToIdTimeout(300.milliseconds)
+        given timeout: FromFutureToIdTimeout =
+          FromFutureToIdTimeout(300.milliseconds)
 
         lazy val fa = Future(a)
         val actual  = FromFuture[Id].toEffect(fa)
