+++ modules/effectie-cats/shared/src/test/scala/effectie/instances/id/fromFutureSpec.scala
@@ -2,10 +2,9 @@ package effectie.instances.id
 
 import cats.Id
 import effectie.core.FromFuture
+import effectie.core.FromFuture.FromFutureToIdTimeout
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
-import fromFuture._
-import effectie.instances.future.fromFuture
 import hedgehog._
 import hedgehog.runner._
 
@@ -27,7 +26,6 @@ object fromFutureSpec extends Properties {
   )
 
   object FutureSpec {
-
     import effectie.instances.future.fromFuture._
 
     def testToEffect: Property = for {
@@ -47,6 +45,8 @@ object fromFutureSpec extends Properties {
   }
 
   object IdSpec {
+    import effectie.instances.id.fromFuture._
+
     def testToEffect: Property = for {
       a <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("a")
     } yield {
@@ -55,8 +55,8 @@ object fromFutureSpec extends Properties {
         ConcurrentSupport.newExecutionContextWithLogger(es, ErrorLogger.printlnExecutionContextErrorLogger)
 
       ConcurrentSupport.runAndShutdown(es, waitFor300Millis) {
-        implicit val timeout: fromFuture.FromFutureToIdTimeout =
-          fromFuture.FromFutureToIdTimeout(waitFor300Millis.waitFor)
+        implicit val timeout: FromFutureToIdTimeout =
+          FromFutureToIdTimeout(waitFor300Millis.waitFor)
 
         lazy val fa = Future(a)
         val actual  = FromFuture[Id].toEffect(fa)
