+++ modules/effectie-core/shared/src/main/scala/effectie/instances/future/fromFuture.scala
@@ -3,7 +3,6 @@ package effectie.instances.future
 import effectie.core.FromFuture
 
 import scala.concurrent.Future
-import scala.concurrent.duration.Duration
 
 /** @author Kevin Lee
   * @since 2021-11-03
@@ -15,7 +14,5 @@ object fromFuture {
       override def toEffect[A](future: => Future[A]): Future[A] =
         future
     }
-
-  final case class FromFutureToIdTimeout(fromFutureToIdTimeout: Duration)
 
 }
