+++ modules/effectie-monix3/shared/src/main/scala/effectie/monix3/fromFuture.scala
@@ -3,7 +3,7 @@ package effectie.monix3
 import cats.Id
 import cats.effect.{ContextShift, IO}
 import effectie.core.FromFuture
-import effectie.instances.future.fromFuture.FromFutureToIdTimeout
+import effectie.core.FromFuture.FromFutureToIdTimeout
 import monix.eval.Task
 
 import scala.concurrent.{Await, Future}
