+++ modules/effectie-syntax/shared/src/main/scala-2/effectie/syntax/console.scala
@@ -1,10 +1,6 @@
 package effectie.syntax
 
-import cats._
-import cats.syntax.all._
-import console.ConsoleEffectF
-import effectie.core.ConsoleEffect.ConsoleEffectWithoutFlatMap
-import effectie.core.{ConsoleEffect, FxCtor, YesNo}
+import effectie.core.{ConsoleEffect, YesNo}
 
 trait console {
 
@@ -22,28 +18,6 @@ trait console {
 
   @inline def readYesNo[F[*]: ConsoleEffect](prompt: String): F[YesNo] = ConsoleEffect[F].readYesNo(prompt)
 
-  implicit def consoleEffectF[F[*]: FxCtor: FlatMap]: ConsoleEffect[F] =
-    new ConsoleEffectF[F]
-
 }
 
-object console extends console {
-
-  final class ConsoleEffectF[F[*]: FxCtor: FlatMap] extends ConsoleEffectWithoutFlatMap[F] with ConsoleEffect[F] {
-
-    @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
-    override def readYesNo(prompt: String): F[YesNo] = for {
-      _      <- putStrLn(prompt)
-      answer <- readLn
-      yesOrN <- answer match {
-                  case "y" | "Y" =>
-                    FxCtor[F].pureOf(YesNo.yes)
-                  case "n" | "N" =>
-                    FxCtor[F].pureOf(YesNo.no)
-                  case _ =>
-                    readYesNo(prompt)
-                }
-    } yield yesOrN
-
-  }
-}
+object console extends console
